Update IAM Group tests
